Refactoring of utils
Introduction of new submodules in utils to break up one huge module.
