+++ src/aspyre/estimation/covar.py
@@ -8,8 +8,9 @@ from tqdm import tqdm
 from aspyre import config
 from aspyre.imaging.threed import rotated_grids
 from aspyre.nfft import anufft3
-from aspyre.utils import mdim_ifftshift, vol_to_vec, vecmat_to_volmat, volmat_to_vecmat, ensure, \
-    symmat_to_vec_iso, vec_to_symmat_iso, make_symmat
+from aspyre.utils.fft import mdim_ifftshift
+from aspyre.utils import ensure
+from aspyre.utils.matrix import vol_to_vec, vecmat_to_volmat, volmat_to_vecmat, symmat_to_vec_iso, vec_to_symmat_iso, make_symmat
 from aspyre.utils.matlab_compat import m_reshape
 from aspyre.estimation import Estimator
 from aspyre.estimation.mean import MeanEstimator
