+++ src/aspyre/estimation/mean.py
@@ -4,7 +4,7 @@ from scipy.fftpack import fft2
 
 from aspyre.imaging.threed import rotated_grids
 from aspyre.nfft import anufft3
-from aspyre.utils import mdim_ifftshift
+from aspyre.utils.fft import mdim_ifftshift
 from aspyre.utils.matlab_compat import m_reshape, m_flatten
 from aspyre.estimation import Estimator
 from aspyre.estimation.kernel import FourierKernel
