+++ src/aspyre/imaging/threed/__init__.py
@@ -1,5 +1,8 @@
 import numpy as np
-from aspyre.utils import grid_2d, centered_ifft2, centered_fft2, ensure
+
+from aspyre.utils import ensure
+from aspyre.utils.math import grid_2d
+from aspyre.utils.fft import centered_ifft2, centered_fft2
 from aspyre.utils.matlab_compat import m_reshape, m_flatten
 from aspyre.nfft import Plan
