+++ tests/test_covar.py
@@ -10,8 +10,9 @@ from aspyre.basis.fb_3d import FBBasis3D
 from aspyre.imaging.filters import RadialCTFFilter
 from aspyre.estimation.mean import MeanEstimator
 from aspyre.estimation.covar import CovarianceEstimator
-from aspyre.utils.mdim import eigs
-from aspyre.utils import src_wiener_coords, Random
+from aspyre.utils.matrix import eigs
+from aspyre.utils.misc import src_wiener_coords
+from aspyre.utils.matlab_compat import Random
 
 import os.path
 DATA_DIR = os.path.join(os.path.dirname(__file__), 'saved_test_data')
