+++ tests/test_random.py
@@ -1,5 +1,5 @@
 from unittest import TestCase
-from aspyre.utils import randi
+from aspyre.utils.matlab_compat import randi
 
 
 class UtilsRandomTestCase(TestCase):
